+++ src/app/curated-rolls/actions.ts
@@ -1,6 +1,6 @@
 import { createStandardAction } from 'typesafe-actions';
-import { CuratedRoll } from './curatedRoll';
+import { CuratedRollsAndInfo } from './curatedRoll';
 
-export const loadCurations = createStandardAction('curations/LOAD')<CuratedRoll[]>();
+export const loadWishLists = createStandardAction('wishlists/LOAD')<CuratedRollsAndInfo>();
 
-export const clearCurations = createStandardAction('curations/CLEAR')();
+export const clearWishLists = createStandardAction('wishlists/CLEAR')();